+++ www/src/components/__tests__/banner.js
@@ -1,6 +1,8 @@
 import React from "react"
 import { render } from "@testing-library/react"
+import { ThemeProvider } from "theme-ui"
 
+import theme from "../../../src/gatsby-plugin-theme-ui"
 import Banner from "../banner"
 
 const matchers = [`Watch now`, `Register now`, `Gatsby Preview`]
@@ -8,7 +10,13 @@ const matchers = [`Watch now`, `Register now`, `Gatsby Preview`]
 const getElement = utils => utils.getByText(text => matchers.includes(text))
 
 test(`it renders an external link`, () => {
-  const link = getElement(render(<Banner />))
+  const link = getElement(
+    render(
+      <ThemeProvider theme={theme}>
+        <Banner />
+      </ThemeProvider>
+    )
+  )
 
   expect(link.getAttribute(`href`)).toContain(`https://`)
 })