+++ www/src/components/layout.js
@@ -1,30 +1,29 @@
+/** @jsx jsx */
+import { jsx } from "theme-ui"
 import React from "react"
 import { navigate, PageRenderer } from "gatsby"
 import mousetrap from "mousetrap"
 import Modal from "react-modal"
-import { SkipNavLink } from "@reach/skip-nav"
 import MdClose from "react-icons/lib/md/close"
 
+import { Global } from "@emotion/core"
+
+import { globalStyles } from "../utils/styles/global"
 import {
   colors,
-  radii,
   space,
-  shadows,
-  sizes,
-  fontSizes,
   zIndices,
-} from "../utils/presets"
+  mediaQueries,
+} from "../gatsby-plugin-theme-ui"
 import { breakpointGutter } from "../utils/styles"
 import Banner from "../components/banner"
+import withColorMode from "../components/with-color-mode"
 import Navigation from "../components/navigation"
 import MobileNavigation from "../components/navigation-mobile"
 import PageWithSidebar from "../components/page-with-sidebar"
 import SiteMetadata from "../components/site-metadata"
-
-// Import Futura PT typeface
+import SkipNavLink from "../components/skip-nav-link"
 import "../assets/fonts/futura"
-
-import { skipLink } from "../utils/styles"
 
 let windowWidth
 
@@ -71,10 +70,12 @@ class DefaultLayout extends React.Component {
     if (this.props.isModal && windowWidth > 750) {
       isModal = true
     }
+    const isDark = this.props.colorMode[0] === `dark`
 
     if (isModal && window.innerWidth > 750) {
       return (
         <>
+          <Global styles={globalStyles} />
           <PageRenderer
             location={{ pathname: this.props.modalBackgroundPath }}
           />
@@ -82,65 +83,81 @@ class DefaultLayout extends React.Component {
             isOpen={true}
             style={{
               content: {
-                top: `inherit`,
-                left: `inherit`,
-                right: `inherit`,
-                bottom: `inherit`,
-                margin: `0 auto`,
-                width: `750px`,
                 background: `none`,
                 border: `none`,
-                padding: `${space[8]} 0`,
+                bottom: `inherit`,
+                left: `inherit`,
+                margin: `0 auto`,
                 overflow: `visible`,
+                padding: `${space[8]} 0`,
+                right: `inherit`,
+                top: `inherit`,
+                maxWidth: `1050px`,
               },
               overlay: {
-                position: `absolute`,
-                top: 0,
-                left: 0,
-                right: 0,
+                backgroundColor: isDark
+                  ? colors.modes.dark.modal.overlayBackground
+                  : colors.modal.overlayBackground,
                 bottom: `unset`,
+                left: 0,
                 minHeight: `100%`,
                 minWidth: `100%`,
-                zIndex: zIndices.modal,
                 overflowY: `auto`,
-                backgroundColor: `rgba(255, 255, 255, 0.95)`,
+                position: `absolute`,
+                right: 0,
+                top: 0,
+                zIndex: zIndices.modal,
               },
             }}
             onRequestClose={() => navigate(this.props.modalBackgroundPath)}
             contentLabel="Site Details Modal"
           >
             <div
-              css={{
-                backgroundColor: colors.white,
-                borderRadius: radii[2],
-                boxShadow: shadows.dialog,
-                position: `relative`,
+              sx={{
+                display: `flex`,
+                flexWrap: `wrap`,
+                justifyContent: `space-between`,
+                [mediaQueries.md]: {
+                  flexWrap: `nowrap`,
+                },
               }}
             >
-              <button
-                onClick={this.handleCloseModal}
-                css={{
-                  background: colors.white,
-                  border: 0,
-                  borderRadius: radii[6],
-                  color: colors.text.secondary,
-                  cursor: `pointer`,
-                  position: `absolute`,
-                  left: `auto`,
-                  right: space[7],
-                  top: space[8],
-                  height: 40,
-                  width: 40,
-                  fontSize: fontSizes[4],
-                  "&:hover": {
-                    background: colors.ui.hover,
-                    color: colors.gatsby,
-                  },
+              <div
+                sx={{
+                  bg: `card.background`,
+                  borderRadius: 2,
+                  boxShadow: `dialog`,
+                  position: `relative`,
+                  alignItems: `center`,
+                  order: 1,
+                  width: `100%`,
                 }}
               >
-                <MdClose />
-              </button>
-              {this.props.children}
+                <button
+                  onClick={this.handleCloseModal}
+                  sx={{
+                    bg: `card.background`,
+                    border: 0,
+                    borderRadius: 6,
+                    color: `textMuted`,
+                    cursor: `pointer`,
+                    fontSize: 4,
+                    height: 40,
+                    left: `auto`,
+                    position: `absolute`,
+                    right: t => t.space[7],
+                    top: t => t.space[8],
+                    width: 40,
+                    "&:hover": {
+                      bg: `ui.hover`,
+                      color: `gatsby`,
+                    },
+                  }}
+                >
+                  <MdClose />
+                </button>
+                {this.props.children}
+              </div>
               {this.props.modalPreviousLink}
               {this.props.modalNextLink}
             </div>
@@ -151,21 +168,22 @@ class DefaultLayout extends React.Component {
 
     return (
       <>
+        <Global styles={globalStyles} />
         <SiteMetadata pathname={this.props.location.pathname} />
-        <SkipNavLink css={skipLink}>Skip to main content</SkipNavLink>
+        <SkipNavLink />
         <Banner />
         <Navigation pathname={this.props.location.pathname} />
         <div
-          className={`main-body`}
-          css={{
-            paddingLeft: `env(safe-area-inset-left)`,
-            paddingRight: `env(safe-area-inset-right)`,
-            paddingTop: sizes.bannerHeight,
+          className={`main-body docSearch-content`}
+          sx={{
+            px: `env(safe-area-inset-left)`,
+            pt: t => t.sizes.bannerHeight,
             // make room for the mobile navigation
-            paddingBottom: sizes.headerHeight,
+            pb: t => t.sizes.headerHeight,
             [breakpointGutter]: {
-              paddingTop: `calc(${sizes.bannerHeight} + ${sizes.headerHeight})`,
-              paddingBottom: 0,
+              pt: t =>
+                `calc(${t.sizes.bannerHeight} + ${t.sizes.headerHeight})`,
+              pb: 0,
             },
           }}
         >
@@ -183,4 +201,4 @@ class DefaultLayout extends React.Component {
   }
 }
 
-export default DefaultLayout
+export default withColorMode(DefaultLayout)