+++ www/src/pages/docs/ssr-apis.js
@@ -1,10 +1,11 @@
+/** @jsx jsx */
+import { jsx } from "theme-ui"
 import React from "react"
 import { graphql } from "gatsby"
 import { Helmet } from "react-helmet"
 import { sortBy } from "lodash-es"
 
 import APIReference from "../../components/api-reference"
-import { space } from "../../utils/presets"
 import Layout from "../../components/layout"
 import Container from "../../components/container"
 import { itemListDocs } from "../../utils/sidebar/item-list"
@@ -41,14 +42,14 @@ class SSRAPIs extends React.Component {
           <h1 id="gatsby-server-rendering-apis" css={{ marginTop: 0 }}>
             Gatsby Server Rendering APIs
           </h1>
-          <h2 css={{ marginBottom: space[3] }}>Usage</h2>
-          <p css={{ marginBottom: space[5] }}>
+          <h2 sx={{ mb: 3 }}>Usage</h2>
+          <p sx={{ mb: 5 }}>
             Implement any of these APIs by exporting them from a file named
             {` `}
             <code>gatsby-ssr.js</code> in the root of your project.
           </p>
           <hr />
-          <h2 css={{ marginBottom: space[3] }}>APIs</h2>
+          <h2 sx={{ mb: 3 }}>APIs</h2>
           <ul>
             {funcs.map((node, i) => (
               <li key={`function list ${node.name}`}>