+++ www/src/pages/newsletter.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react"
 import { Helmet } from "react-helmet"
-import { colors, space } from "../utils/presets"
 
 import Layout from "../components/layout"
 import Container from "../components/container"
@@ -39,11 +38,6 @@ class NewsLetter extends Component {
           <EmailCaptureForm
             signupMessage="Sign up for the Gatsby Newsletter"
             confirmMessage="Success! You have been subscribed to the Gatsby newsletter. Expect to see a newsletter in your inbox each Wednesday (or the equivalent of US Wednesday in your time zone)!"
-            overrideCSS={{
-              marginTop: space[5],
-              paddingTop: space[3],
-              borderTop: `2px solid ${colors.lilac}`,
-            }}
           />
         </Container>
         <FooterLinks />