+++ www/src/views/showcase/collapsible-filter-list.js
@@ -1,11 +1,10 @@
-import React from "react"
+/** @jsx jsx */
+import { jsx } from "theme-ui"
 
 import MdCheckboxBlank from "react-icons/lib/md/check-box-outline-blank"
 import MdCheckbox from "react-icons/lib/md/check-box"
 import Collapsible from "../shared/collapsible"
-
-import { colors } from "../../utils/presets"
-import styles from "../../views/shared/styles"
+import { filterButton, filterCheckbox } from "../../views/shared/styles"
 
 const CollapsibleFilterList = ({
   filters,
@@ -26,29 +25,20 @@ const CollapsibleFilterList = ({
             setFilters([...filters, c])
           }
         }}
-        css={styles.filterButton}
+        sx={{
+          ...filterButton,
+          color: filters.includes(c) ? `link.color` : `textMuted`,
+        }}
       >
-        <div
-          css={{
-            color: filters.includes(c) ? colors.gatsby : colors.input.border,
-            ...styles.filterCheckbox,
-          }}
-        >
+        <div sx={filterCheckbox}>
           {filters.includes(c) ? (
             <MdCheckbox style={{ verticalAlign: `-0.125em` }} />
           ) : (
             <MdCheckboxBlank style={{ verticalAlign: `-0.125em` }} />
           )}
         </div>
-        <div
-          css={{
-            color: filters.includes(c) ? colors.gatsby : false,
-            marginRight: `auto`,
-          }}
-        >
-          {c}
-        </div>
-        <div css={styles.filterCount}>{aggregatedCategories[c]}</div>
+        <div sx={{ mr: `auto` }}>{c}</div>
+        <div>{aggregatedCategories[c]}</div>
       </button>
     ))}
   </Collapsible>