+++ packages/gatsby-core-utils/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.17...gatsby-core-utils@1.0.18) (2019-11-10)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.16...gatsby-core-utils@1.0.17) (2019-10-28)
+
+### Features
+
+- **gatsby-core-utils:** Add isCI and getCIName ([#19039](https://github.com/gatsbyjs/gatsby/issues/19039)) ([d97bee6](https://github.com/gatsbyjs/gatsby/commit/d97bee6))
+
+## [1.0.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.15...gatsby-core-utils@1.0.16) (2019-10-28)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.14...gatsby-core-utils@1.0.15) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.13...gatsby-core-utils@1.0.14) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.12...gatsby-core-utils@1.0.13) (2019-10-09)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.10...gatsby-core-utils@1.0.12) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.10...gatsby-core-utils@1.0.11) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.9...gatsby-core-utils@1.0.10) (2019-09-20)
+
+**Note:** Version bump only for package gatsby-core-utils
+
+## [1.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.8...gatsby-core-utils@1.0.9) (2019-09-18)
+
+**Note:** Version bump only for package gatsby-core-utils
+
 ## [1.0.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-core-utils@1.0.7...gatsby-core-utils@1.0.8) (2019-09-09)
 
 **Note:** Version bump only for package gatsby-core-utils