+++ packages/gatsby-plugin-sharp/CHANGELOG.md
@@ -3,6 +3,92 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.38](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.37...gatsby-plugin-sharp@2.2.38) (2019-11-13)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.2.37](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.36...gatsby-plugin-sharp@2.2.37) (2019-11-10)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.2.36](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.35...gatsby-plugin-sharp@2.2.36) (2019-10-29)
+
+### Bug Fixes
+
+- **gatsby-plugin-sharp:** fix "Cannot set property 'multipassCount' of undefined" error ([#19131](https://github.com/gatsbyjs/gatsby/issues/19131)) ([3766dcd](https://github.com/gatsbyjs/gatsby/commit/3766dcd))
+
+## [2.2.35](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.34...gatsby-plugin-sharp@2.2.35) (2019-10-29)
+
+### Bug Fixes
+
+- update minor updates in packages ([#19087](https://github.com/gatsbyjs/gatsby/issues/19087)) ([294a48f](https://github.com/gatsbyjs/gatsby/commit/294a48f))
+
+## [2.2.34](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.33...gatsby-plugin-sharp@2.2.34) (2019-10-28)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.2.33](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.32...gatsby-plugin-sharp@2.2.33) (2019-10-28)
+
+### Bug Fixes
+
+- update minor updates in packages ([#18875](https://github.com/gatsbyjs/gatsby/issues/18875)) ([b692879](https://github.com/gatsbyjs/gatsby/commit/b692879))
+
+## [2.2.32](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.31...gatsby-plugin-sharp@2.2.32) (2019-10-16)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.2.31](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.30...gatsby-plugin-sharp@2.2.31) (2019-10-14)
+
+### Bug Fixes
+
+- **gatsby-plugin-sharp:** Allow brackets in paths ([#18289](https://github.com/gatsbyjs/gatsby/issues/18289)) ([f9933b1](https://github.com/gatsbyjs/gatsby/commit/f9933b1))
+
+## [2.2.30](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.29...gatsby-plugin-sharp@2.2.30) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.2.29](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.28...gatsby-plugin-sharp@2.2.29) (2019-10-09)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.2.28](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.27...gatsby-plugin-sharp@2.2.28) (2019-10-04)
+
+### Bug Fixes
+
+- update dependency bluebird to ^3.7.0 ([#18029](https://github.com/gatsbyjs/gatsby/issues/18029)) ([bd235a8](https://github.com/gatsbyjs/gatsby/commit/bd235a8))
+
+## [2.2.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.25...gatsby-plugin-sharp@2.2.27) (2019-09-26)
+
+### Bug Fixes
+
+- update dependency sharp to ^0.23.1 ([#17904](https://github.com/gatsbyjs/gatsby/issues/17904)) ([f19ad3c](https://github.com/gatsbyjs/gatsby/commit/f19ad3c))
+
+## [2.2.26](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.25...gatsby-plugin-sharp@2.2.26) (2019-09-26)
+
+### Bug Fixes
+
+- update dependency sharp to ^0.23.1 ([#17904](https://github.com/gatsbyjs/gatsby/issues/17904)) ([f19ad3c](https://github.com/gatsbyjs/gatsby/commit/f19ad3c))
+
+## [2.2.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.24...gatsby-plugin-sharp@2.2.25) (2019-09-20)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.2.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.23...gatsby-plugin-sharp@2.2.24) (2019-09-18)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.2.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.22...gatsby-plugin-sharp@2.2.23) (2019-09-18)
+
+### Bug Fixes
+
+- **gatsby-plugin-sharp:** Set a unique uuid for each job ([#17693](https://github.com/gatsbyjs/gatsby/issues/17693)) ([0926c73](https://github.com/gatsbyjs/gatsby/commit/0926c73))
+
+## [2.2.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.21...gatsby-plugin-sharp@2.2.22) (2019-09-13)
+
+### Features
+
+- **gatsby-remark-images:** Add flag to suppress css background-image to prevent FOUB ([#17154](https://github.com/gatsbyjs/gatsby/issues/17154)) ([125fd01](https://github.com/gatsbyjs/gatsby/commit/125fd01))
+
 ## [2.2.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.2.20...gatsby-plugin-sharp@2.2.21) (2019-09-09)
 
 **Note:** Version bump only for package gatsby-plugin-sharp