+++ packages/gatsby-remark-smartypants/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-smartypants@2.1.14...gatsby-remark-smartypants@2.1.15) (2019-11-10)
+
+**Note:** Version bump only for package gatsby-remark-smartypants
+
+## [2.1.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-smartypants@2.1.13...gatsby-remark-smartypants@2.1.14) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-remark-smartypants
+
+## [2.1.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-smartypants@2.1.12...gatsby-remark-smartypants@2.1.13) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-remark-smartypants
+
+## [2.1.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-smartypants@2.1.11...gatsby-remark-smartypants@2.1.12) (2019-10-09)
+
+**Note:** Version bump only for package gatsby-remark-smartypants
+
+## [2.1.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-smartypants@2.1.9...gatsby-remark-smartypants@2.1.11) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-remark-smartypants
+
+## [2.1.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-smartypants@2.1.9...gatsby-remark-smartypants@2.1.10) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-remark-smartypants
+
+## [2.1.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-smartypants@2.1.8...gatsby-remark-smartypants@2.1.9) (2019-09-20)
+
+**Note:** Version bump only for package gatsby-remark-smartypants
+
 ## [2.1.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-smartypants@2.1.7...gatsby-remark-smartypants@2.1.8) (2019-09-09)
 
 **Note:** Version bump only for package gatsby-remark-smartypants