+++ packages/gatsby-transformer-documentationjs/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-documentationjs@4.1.16...gatsby-transformer-documentationjs@4.1.17) (2019-11-10)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
+## [4.1.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-documentationjs@4.1.15...gatsby-transformer-documentationjs@4.1.16) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
+## [4.1.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-documentationjs@4.1.14...gatsby-transformer-documentationjs@4.1.15) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
+## [4.1.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-documentationjs@4.1.13...gatsby-transformer-documentationjs@4.1.14) (2019-10-09)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
+## [4.1.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-documentationjs@4.1.11...gatsby-transformer-documentationjs@4.1.13) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
+## [4.1.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-documentationjs@4.1.11...gatsby-transformer-documentationjs@4.1.12) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
+## [4.1.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-documentationjs@4.1.10...gatsby-transformer-documentationjs@4.1.11) (2019-09-20)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
 ## [4.1.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-documentationjs@4.1.9...gatsby-transformer-documentationjs@4.1.10) (2019-09-10)
 
 **Note:** Version bump only for package gatsby-transformer-documentationjs